Comite das mudanças de autenticação e autorização.

+++ SamaCardAll.Core/Interfaces/IUserRepository.cs
@@ -5,7 +5,7 @@ namespace SamaCardAll.Core.Interfaces
     public interface IUserRepository
     {
         Task<User?> GetByEmailAsync(string email);
-        Task<long> CreateAsync(User user);  
-        Task<User?> GetByIdAsync(long id);
+        Task<int> CreateAsync(User user);  
+        Task<User?> GetByIdAsync(int id);
     }
 }
